+++ Realm/Realm.Sync/User.cs
@@ -17,6 +17,7 @@
 ////////////////////////////////////////////////////////////////////////////
 
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Realms.Exceptions;
@@ -74,23 +75,23 @@ namespace Realms.Sync
         /// Logs the user in to the Realm Object Server.
         /// </summary>
         /// <param name="credentials">The credentials to use for authentication.</param>
-        /// <param name="serverUrl">The URI of the server that the user is authenticated against.</param>
+        /// <param name="serverUri">The URI of the server that the user is authenticated against.</param>
         /// <returns>An awaitable Task, that, upon completion, contains the logged in user.</returns>
-        public static async Task<User> LoginAsync(Credentials credentials, Uri serverUrl)
+        public static async Task<User> LoginAsync(Credentials credentials, Uri serverUri)
         {
             Argument.NotNull(credentials, nameof(credentials));
-            Argument.NotNull(serverUrl, nameof(serverUrl));
-            Argument.Ensure(serverUrl.Scheme.StartsWith("http"), "Unexpected protocol for login url. Expected http:// or https://.", nameof(serverUrl));
+            Argument.NotNull(serverUri, nameof(serverUri));
+            Argument.Ensure(serverUri.Scheme.StartsWith("http"), "Unexpected protocol for login url. Expected http:// or https://.", nameof(serverUri));
 
             SharedRealmHandleExtensions.DoInitialFileSystemConfiguration();
 
             if (credentials.IdentityProvider == Credentials.Provider.AdminToken)
             {
-                return new User(SyncUserHandle.GetAdminTokenUser(serverUrl.AbsoluteUri, credentials.Token));
+                return new User(SyncUserHandle.GetAdminTokenUser(serverUri.AbsoluteUri, credentials.Token));
             }
 
-            var result = await AuthenticationHelper.LoginAsync(credentials, serverUrl);
-            var handle = SyncUserHandle.GetSyncUser(result.UserId, serverUrl.AbsoluteUri, result.RefreshToken, result.IsAdmin);
+            var result = await AuthenticationHelper.LoginAsync(credentials, serverUri);
+            var handle = SyncUserHandle.GetSyncUser(result.UserId, serverUri.AbsoluteUri, result.RefreshToken, result.IsAdmin);
             return new User(handle);
         }
 
@@ -124,12 +125,12 @@ namespace Realms.Sync
         /// </summary>
         /// <returns>A user instance if a logged in user with that id exists, <c>null</c> otherwise.</returns>
         /// <param name="identity">The identity of the user.</param>
-        /// <param name="serverUrl">The URI of the server that the user is authenticated against.</param>
-        public static User GetLoggedInUser(string identity, Uri serverUrl)
+        /// <param name="serverUri">The URI of the server that the user is authenticated against.</param>
+        public static User GetLoggedInUser(string identity, Uri serverUri)
         {
             SharedRealmHandleExtensions.DoInitialFileSystemConfiguration();
 
-            if (SyncUserHandle.TryGetLoggedInUser(identity, serverUrl.AbsoluteUri, out var userHandle))
+            if (SyncUserHandle.TryGetLoggedInUser(identity, serverUri.AbsoluteUri, out var userHandle))
             {
                 return new User(userHandle);
             }
@@ -305,6 +306,94 @@ namespace Realms.Sync
             Argument.NotNullOrEmpty(providerUserIdentity, nameof(providerUserIdentity));
 
             return AuthenticationHelper.RetrieveInfoForUserAsync(this, provider, providerUserIdentity);
+        }
+
+        /// <summary>
+        /// Request a password reset email to be sent to a user's email. This method requires internet connection
+        /// and will not throw an exception, even if the email doesn't belong to a Realm Object Server user.
+        /// </summary>
+        /// <remarks>
+        /// This can only be used for users who authenticated with <see cref="Credentials.UsernamePassword"/>
+        /// and passed a valid email address as a username.
+        /// </remarks>
+        /// <param name="serverUri">The URI of the server that the user is authenticated against.</param>
+        /// <param name="email">The email that corresponds to the user's username.</param>
+        /// <returns>An awaitable task that, upon completion, indicates that the request has been sent.</returns>
+        public static Task RequestPasswordResetAsync(Uri serverUri, string email)
+        {
+            Argument.NotNull(serverUri, nameof(serverUri));
+            Argument.NotNullOrEmpty(email, nameof(email));
+
+            return AuthenticationHelper.UpdateAccountAsync(serverUri, "reset_password", email);
+        }
+
+        /// <summary>
+        /// Complete the password reset flow by using the reset token sent to the user's email as a one-time
+        /// authorization token to change the password.
+        /// </summary>
+        /// <remarks>
+        /// By default, the link that will be sent to the user's email will redirect to a webpage where
+        /// they can enter their new password. If you wish to provide a native UX, you may wish to modify
+        /// the url to use deep linking to open the app, extract the token, and navigate to a view that
+        /// allows them to change their password within the app.
+        /// </remarks>
+        /// <param name="serverUri">The URI of the server that the user is authenticated against.</param>
+        /// <param name="token">The token that was sent to the user's email address.</param>
+        /// <param name="newPassword">The user's new password.</param>
+        /// <returns>An awaitable task that, when successful, indicates that the password has changed.</returns>
+        public static Task CompletePasswordResetAsync(Uri serverUri, string token, string newPassword)
+        {
+            Argument.NotNull(serverUri, nameof(serverUri));
+            Argument.NotNullOrEmpty(token, nameof(token));
+            Argument.NotNullOrEmpty(newPassword, nameof(newPassword));
+
+            var data = new Dictionary<string, string>
+            {
+                ["token"] = token,
+                ["new_password"] = newPassword
+            };
+            return AuthenticationHelper.UpdateAccountAsync(serverUri, "complete_reset", data: data);
+        }
+
+        /// <summary>
+        /// Request an email confirmation email to be sent to a user's email. This method requires internet connection
+        /// and will not throw an exception, even if the email doesn't belong to a Realm Object Server user.
+        /// </summary>
+        /// <param name="serverUri">The URI of the server that the user is authenticated against.</param>
+        /// <param name="email">The email that corresponds to the user's username.</param>
+        /// <returns>An awaitable task that, upon completion, indicates that the request has been sent.</returns>
+        public static Task RequestEmailConfirmationAsync(Uri serverUri, string email)
+        {
+            Argument.NotNull(serverUri, nameof(serverUri));
+            Argument.NotNullOrEmpty(email, nameof(email));
+
+            return AuthenticationHelper.UpdateAccountAsync(serverUri, "request_email_confirmation", email);
+        }
+
+        /// <summary>
+        /// Complete the password reset flow by using the confirmation token sent to the user's email as a one-time
+        /// authorization token to confirm their email.
+        /// </summary>
+        /// <remarks>
+        /// By default, the link that will be sent to the user's email will redirect to a webpage where
+        /// they'll see a generic "Thank you for confirming" text. If you wish to provide a native UX, you
+        /// may wish to modify the url to use deep linking to open the app, extract the token, and inform them
+        /// that their email has been confirmed.
+        /// </remarks>
+        /// <param name="serverUri">The URI of the server that the user is authenticated against.</param>
+        /// <param name="token">The token that was sent to the user's email address.</param>
+        /// <returns>An awaitable task that, when successful, indicates that the email has been confirmed.</returns>
+        public static Task ConfirmEmailAsync(Uri serverUri, string token)
+        {
+            Argument.NotNull(serverUri, nameof(serverUri));
+            Argument.NotNullOrEmpty(token, nameof(token));
+
+            var data = new Dictionary<string, string>
+            {
+                ["token"] = token
+            };
+
+            return AuthenticationHelper.UpdateAccountAsync(serverUri, "confirm_email", data: data);
         }
 
         /// <inheritdoc />